+++ lib/common/globals.h
@@ -85,7 +85,8 @@ GLOBALS_API EXTERN Agsym_t *E_weight, *E_minlen, *E_color, *E_fillcolor,
     *E_style, *E_decorate, *E_showboxes, *E_arrowsz, *E_constr, *E_layer,
     *E_comment, *E_label_float, *E_samehead, *E_sametail, *E_headlabel,
     *E_taillabel, *E_labelfontsize, *E_labelfontname, *E_labelfontcolor,
-    *E_labeldistance, *E_labelangle, *E_tailclip, *E_headclip, *E_penwidth;
+    *E_labeldistance, *E_labelangle, *E_tailclip, *E_headclip, *E_penwidth,
+    *E_midarrowhead, *E_midarrowtail, *E_midarrowpos;
 
 GLOBALS_API extern struct fdpParms_s *fdp_parms;