+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to
 
 ## [Unreleased (16.1.1)]
 
+### Added
+
+- Edges can draw optional middle arrowheads and arrowtails with the
+  `midarrowhead`, `midarrowtail`, and `midarrowpos` attributes. #2005
+
 ### Fixed
 
 - `nop` no longer reorders edges. #2855